Fix map and price

+++ app/src/main/java/ru/lionzxy/printbox/view/print_history/ui/HistoryAdapter.kt
@@ -26,7 +26,7 @@ class HistoryAdapter(var history: List<PrintHistory>,
         } else {
             holder.orderStatus.setBackgroundResource(R.drawable.circle_red)
         }
-        holder.orderPrice.text = holder.itemView.context.resources.getString(R.string.history_item_price, (element.price / 100).toFloat())
+        holder.orderPrice.text = holder.itemView.context.resources.getString(R.string.history_item_price, (element.price.toDouble() / 100).toFloat())
         holder.orderFilename.text = element.document.name
         holder.orderFiledate.text = element.createdAt.toLocalDateTime().toString("yyyy-MM-dd HH:mm")
         holder.orderTextStatus.text = element.printerName
